tests: refactor luaL_loadbuffer_proto
This refactoring will:

1. Move macros from a header to the source file.
Macros should be used in header only with undef to avoid redefinitions.
Undef directive is not useful since we want to use these macros in the
source file.

2. Remove `using namespace lua_grammar` from header [1].

3. Moving serializer entry point and constant parameters into
luajit_fuzzer namespace.
It's a common practice in C++ to avoid name collisions.

4. Move serializer functions into anonymous namespace.
These functions are not a part of the interface so should have
static linkage [2].

5. Fix ConvertToStringDefault function.
It was logically wrong so it would generate an identifier `123` from
`*123`.

1. https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rs-using-directive
2. https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rs-unnamed2

(cherry-picked from commit 56488e15cc697164bbad2541f8e9f26f38964200)