tests: print metrics on sending SIGUSR1 to a test process
By default test luaL_loadbuffer_proto_test prints metrics only when
running with a finite number of runs or finite duration of time (options
`-runs` and `-max_total_time`). It is useful to know metrics when test
is running in endless mode.

Path adds a signal handler for SIGUSR1 that prints metrics on sending
USR1 to a test process (`kill -SIGUSR1 <pid>`).