cmake: don't propagate CFLAGS to LibProtobufMutator
Fuzz Introspector reports refers to a huge number of functions that are
not a target for fuzzing testing and thus are not interesting: Protobuf
library and LibProtobufMutator library.

OSS Fuzz infrastructure set CFLAGS and CXXFLAGS via environment
variables before running building, these compiler flags contain
instrumentation flags and propagated to aforementioned libraries. The
patch fixes that.