cmake: pass a number of runs via environment variable
Commit 121fd6f371da ("tests: reduce a number of test runs") reduced
default number of test runs. However, it is desired to control a number
of testruns with local testing.

The patch adds support of environment variable `RUNS` that is used by
CMake and CTest and passed to test executables:

$ RUNS=10 ctest -R luaL_gsub_test --verbose
<snipped>
1: Done 10 runs in 0 second(s)

By default the same number of runs is used:

$ ctest -R luaL_gsub_test --verbose
<snipped>
1: Done 5 runs in 0 second(s)

(It is not trivial to pass command line arguments in CTest at runtime,
see [1] and [2].)

1. https://stackoverflow.com/questions/28812533/how-to-pass-command-line-arguments-in-ctest-at-runtime
2. https://gitlab.kitware.com/cmake/cmake/-/issues/20470