tests: execute bytecode in luaL_loadbuffer_proto_test
Lua always precompiles source code to an intermediate form before
running it. Moreover, Lua API has a functions that compile Lua code to
bytecode - `loadstring` [1], and dump resulted bytecode to a string -
`string.dump` [2][3]. Results of executed bytecode and Lua code must be
equal:

```
> f = assert(loadstring('return 14')); f() -- 14
> loadstring(string.dump(f))()             -- 14
```

The patch introduces a function `luaL_loadbytecode` that has the same
arguments as a function `luaL_loadbuffer`, but loads a buffer as a
bytecode chunk instead of a Lua chunk to the Lua stack. The test
`luaL_loadbuffer_proto_test` uses function `luaL_loadbytecode` for
loading bytecode, and then executes it.

1. https://www.lua.org/manual/5.1/manual.html#pdf-loadstring
2. https://www.lua.org/manual/5.1/manual.html#pdf-string.dump
3. https://luajit.org/extensions.html
4. https://www.lua.org/pil/8.html