cfl: fix randomly failed builds
GitHub action builds started to randomly fail with a reason:
"The runner has received a shutdown signal.".
The issue [1] and discussion [2] points to a workaround:
replace a "ubuntu-latest" with "ubuntu-20.04" and it started to
work. Note that the problem reproduce with "ubuntu-22.04" as well.

The patch replaces "ubuntu-latest" with "ubuntu-20.04" in CFL
workflows.

1. https://github.com/actions/runner-images/issues/6709
2. https://github.com/actions/runner-images/discussions/7188