cmake: remove arguments in endif()
The arguments to `else()` and `endif()` were required before
version 2.6.0. As of CMake 2.6.0 the `else()` and `endif()`
constructs can be empty. The same is true for closing constructs
on `endmacro()`, `endfunction()`, and `endforeach()`. If you
require 2.4.x compatibility, CMake 2.4.3 or greater recognizes the
`CMAKE_ALLOW_LOOSE_LOOP_CONSTRUCTS` option (which is superfluous
in 2.6.0). See [1] and [2].

The patch removes arguments in `endif()`.

1. https://cmake.org/cmake/help/v3.6/command/endif.html
2. https://stackoverflow.com/questions/29959126/why-does-cmake-syntax-have-redundant-parentheses-everywhere/29967305#29967305