tests/lapi: add string.buffer tests
The patch add tests for LuaJIT's string buffer library [1].

Note, as it is stated in documentation [1] this serialization
format is designed for internal use by LuaJIT applications, and
this format is explicitly not intended to be a 'public standard'
for structured data interchange across computer languages (like
JSON or MessagePack). The purpose of the proposed tests is testing
the library because other LuaJIT components relies on it and
also the proposed tests indirectly tests FFI library.

1. https://luajit.org/ext_buffer.html