tests/lapi: fix assertion with result in os_time_test
The `mktime()` function (which is used inside `os.time()`)
returns -1 for cases when the time since the Epoch cannot be
represented. In this case, `os.time()` returns `nil`:

```
luajit -e "print(os.time({year = 1970, month = 1, day = 1, sec = -9 * 60 * 60 - 1}))"
nil
```

The patch fixes an assertion in the test and allows `nil` as
a result of the function.

Follows up #141